requirements.txt
# RAG Pipeline
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
faiss-cpu>=1.9.0

# Backend API
fastapi>=0.115.0
uvicorn>=0.32.0

# Frontend
streamlit>=1.40.0

# Export
python-docx>=1.1.0
fpdf2>=2.8.0
markdown>=3.7

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
bcrypt>=4.0.0
